requirements.txt
fastapi==0.115.12
python-multipart==0.0.20
uvicorn==0.33.0
python-dotenv ==1.0.1
pydantic_settings==2.8.1
aiofiles == 24.1.0
langchain==0.2.17
PyMuPDF ==1.24.11
langchain-community==0.2.19
motor==3.6.1
# black==25.1.0
# DateTime==5.5
# keras==3.9.0
# matplotlib==3.10.1
# numpy==2.1.3
# pandas==2.2.3
# pylint==3.3.5
# pytest==8.3.5
# scikit-learn==1.6.1
# scipy==1.15.2
# seaborn==0.13.2
# statsmodels==0.14.4
# tensorflow==2.19.0
# lightgbm==4.6.0
# coverage == 7.7.0
# fire==0.7.0
# flask==3.1.0
# flasgger
# flask_jwt_extended
# docker==7.1.0